several ways to use RTT
